+++ libraries/HID/HID.h
@@ -129,9 +129,9 @@ public:
   int begin(void);
 };
 
-int   HID_Plug(void);
-int   HID_GetInterface(u8* interfaceNum);
-int   HID_GetDescriptor(int t);
+int8_t   HID_Plug(void);
+int8_t   HID_GetInterface(u8* interfaceNum);
+int8_t   HID_GetDescriptor(int8_t t);
 bool  HID_Setup(Setup& setup, u8 i);
 void  HID_SendReport(uint8_t id, const void* data, int len);
